Refactor library base.
