+++ README
@@ -1,0 +1,4 @@
+# Asuka 
+
+A statically typed Discord API wrapper using python.
+
