+++ asuka/builders.py
@@ -1,3 +1,27 @@
+# MIT License
+
+# Copyright (c) 2022 Sarthak
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+from __future__ import annotations
+
 import datetime
 import typing
 
@@ -31,39 +55,22 @@ class DiscordObject:
 @typing.final
 class Intents:
     NONE = 0
-
     GUILDS = 1 << 0
-
     GUILD_MEMBERS = 1 << 1
-
     GUILD_BANS = 1 << 2
-
     GUILD_EMOJIS = 1 << 3
-
     GUILD_INTEGRATIONS = 1 << 4
-
     GUILD_WEBHOOKS = 1 << 5
-
     GUILD_INVITES = 1 << 6
-
     GUILD_VOICE_STATES = 1 << 7
-
     GUILD_PRESENCES = 1 << 8
-
     GUILD_MESSAGES = 1 << 9
-
     GUILD_MESSAGE_REACTIONS = 1 << 10
-
     GUILD_MESSAGE_TYPING = 1 << 11
-
     DM_MESSAGES = 1 << 12
-
     DM_MESSAGE_REACTIONS = 1 << 13
-
     DM_MESSAGE_TYPING = 1 << 14
-
     MESSAGE_CONTENT = 1 << 15
-
     GUILD_SCHEDULED_EVENTS = 1 << 16
 
     def __init__(
