+++ asuka/event_handler.py
@@ -1,10 +1,34 @@
+# MIT License
+
+# Copyright (c) 2022 Sarthak
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+from __future__ import annotations
+
 import asyncio
 import typing
 
-from .base_events import GatewayEvent
+from .events.base_events import GatewayEvent
 
 if typing.TYPE_CHECKING:
-    from clyde.bot import Bot
+    from asuka.bot import Bot
 
 
 class EventHandler:
@@ -43,7 +67,9 @@ class EventHandler:
             self.once_listeners[event_class] = [callback]
 
     def dispatch(
-        self, event_type: typing.Any, payload: typing.Dict[typing.Any, typing.Any]
+        self,
+        event_type: typing.Type[GatewayEvent],
+        payload: typing.Dict[typing.Any, typing.Any],
     ) -> None:
         event = event_type(self.bot, payload)
         to_call = []
@@ -52,5 +78,6 @@ class EventHandler:
         )
         for listener in self.once_listeners.get(event_type, []):
             to_call.append(listener(event))
+        self.once_listeners[event_type] = []
 
         asyncio.gather(*to_call)
