+++ asuka/gateway/keep_alive.py
@@ -1,0 +1,45 @@
+# MIT License
+
+# Copyright (c) 2022 Sarthak
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+from __future__ import annotations
+
+import asyncio
+import time
+import typing
+
+from .enums import WSEventEnums
+
+if typing.TYPE_CHECKING:
+    from .gateway import Gateway
+
+
+class KeepAlive:
+    sequence: int = 0
+    last_heartbeat: float
+
+    async def start(self, ws: "Gateway") -> None:
+        while True:
+            await ws.socket.send_json(
+                {"op": WSEventEnums.HEARTBEAT, "d": self.sequence}
+            )
+            self.last_heartbeat = time.perf_counter()
+            await asyncio.sleep(ws._heartbeat_interval)
