--- clyde/__init__.py
@@ -1,3 +1,0 @@
-from .bot import *
-from .builders import *
-from .events import *
