--- clyde/events/__init__.py
@@ -1,3 +1,0 @@
-from .base_events import *
-from .event_handler import *
-from .messages import *
