--- clyde/users.py
@@ -1,5 +1,0 @@
-import typing
-
-
-class PartialUser:
-    ...
