--- clyde/websocket/__init__.py
@@ -1,1 +1,0 @@
-from .ws import DiscordWebSocket
