--- clyde/websocket/enums.py
@@ -1,17 +1,0 @@
-import dataclasses
-import enum
-
-
-@dataclasses.dataclass
-class WSEventEnums(enum.IntFlag):
-    DISPATCH = 0
-    HEARTBEAT = 1
-    IDENTIFY = 2
-    PRESECNE_UPDATE = 3
-    VOICE_STATE_UPDATE = 4
-    RESUME = 6
-    RECONNECT = 7
-    REQUEST_GUILD_MEMBERS = 8
-    INVALID_SESSION = 9
-    HELLO = 10
-    HEARTBEAT_ACK = 11
