+++ pyproject.toml
@@ -1,0 +1,19 @@
+[tool.poetry]
+name = "Asuka"
+version = "0.0.1"
+description = ""
+authors = ["sarthhh"]
+license = "MIT License"
+
+[tool.poetry.dependencies]
+python = "^3.8"
+aiohttp = "^3.8.1"
+
+[tool.poetry.dev-dependencies]
+mypy = "^0.971"
+isort = "^5.10.1"
+black = "^22.6.0"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
